requirements.txt
Flask==2.3.2
requests>=2.30.0
numpy>=1.25.0
Pillow>=10.0.0
gunicorn==20.1.0